Feat: Add ArticleController
* ArticleController를 추가하여 ApiController에서 분리
* getArticle 메소드 추가

-
